Synth_update (READY-TO-MERGE) (#321)
* Alarm (#7)

* Synth_update (WIP)

* Отключение кучи реагентов для синтетиков

* Синтетик ибольше не гниют

* Урон токсинами вместо бёрна по лежащим синтетикам

* Синтетики не могут быть хаскнуты

* Suit Cooling Device

* Даём синтетикам защиту от низкой температуры за бортом (они будут перегреваться)

* Убираем привычный хладоген у синтетиков

* Повышаем значение, при котором невозможно производить саморемонт дальше (с 10 до 20 урона)

* Добавляем отчистку от радиации в операцию отчистки от загрязнений

* Фикс токс урона в крите

* Фикс компила

* Защита глаз от сварки для синтетиков

* Дополнение к защите от сварки

* Фикс рантаймов радиации у РБМК (не связано с ПР)

* Переделка охлаждающего девайса

* Охлаждающий девайс в базовые технологии

* Охлаждающий девайс шахтёрам

* Делаем лаваленд неприятным для синтетиков

* Убираем резисты к холоду

* Убираем защиту от атмосферы за ношение скафандра или костюма на лаваленде

* Защита от вирусов

* Лишнее изменение

* Лишнее изменение

* Документация

* Фикс спавна ПОУ у шахтёров

* ПОУ на слот брони

* Документация + убираем лишнее

* Кулеры в скафандры

* Синты получаем корапшн вместо брута при остановке сердца

* Возможность органов ИПК полностью отказать от ЭМИ

* emp_act для ПОУ и typo

* Запрещаем квирк кровопийцы для синтетиков (ломает способ питания через энергию)

* Выдаём тарконцам ПОУ при спавне за синтетика

* Фикс зарядки батарейки через ПОУ

* Описание для зарядки.

* Typo

* Typo

* Typo

* Порог минимального урона и повышенный урон с руки для синтетиков

* Фикс рантаймов из-за луж отходов РБМК

* Возвращаем старые значения порога урона

* ПОУ на броню СБ

* Убираем клон-урон с роботов

* Убираем воздействие некоторых химикатов на синтетиков

* Выдаём ПОУ вместо одежды тарконцев в их руки (чтобы синты могли брать пальто и другие вещи из лодаута там)

* Typo и документация

+++ code/modules/vore/eating/bellymodes.dm
@@ -226,8 +226,9 @@
 				// Deal digestion damage (and feed the pred)
 				if(!(M.status_flags & GODMODE))
 					M.adjustFireLoss(digest_burn)
-					M.adjustToxLoss(2) // something something plasma based acids
-					M.adjustCloneLoss(1) // eventually this'll kill you if you're healing everything else, you nerds.
+					M.adjustToxLoss(2, toxins_type = TOX_OMNI) // something something plasma based acids // BLUEMOON CHANGES - TOX_OMNI для урона системам синтетика
+					if(!HAS_TRAIT(M, TRAIT_ROBOTIC_ORGANISM)) // BLUEMOON ADD START - у роботов не должно быть клон-урона
+						M.adjustCloneLoss(1) // eventually this'll kill you if you're healing everything else, you nerds.
 				//Contaminate or gurgle items
 			var/obj/item/T = pick(touchable_items)
 			if(istype(T))
